Add a proper check for OnEvent.Executor.Immediate for Node.js
