+++ library/runtime-core/src/nonJsMain/kotlin/io/matthewnelson/kmp/tor/runtime/core/internal/-NonJsOnEventExecutors.kt
@@ -20,10 +20,35 @@ package io.matthewnelson.kmp.tor.runtime.core.internal
 import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
 import io.matthewnelson.kmp.tor.runtime.core.Executable
 import io.matthewnelson.kmp.tor.runtime.core.OnEvent
+import kotlinx.coroutines.*
 import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
 
-internal expect object ExecutorMainInternal: OnEvent.Executor {
+internal actual object ExecutorMainInternal: OnEvent.Executor {
+
+    private val MainScope by lazy {
+        val mainDispatcher = run {
+            // Will throw if Missing
+            Dispatchers.Main.isDispatchNeeded(EmptyCoroutineContext)
+
+            try {
+                Dispatchers.Main.immediate
+            } catch (_: UnsupportedOperationException) {
+                Dispatchers.Main
+            }
+        }
+
+        CoroutineScope(context =
+            CoroutineName("OnEvent.Executor.Main")
+            + SupervisorJob()
+            + mainDispatcher
+        )
+    }
 
     @InternalKmpTorApi
-    override fun execute(handler: CoroutineContext, executable: Executable)
+    actual override fun execute(handler: CoroutineContext, executable: Executable) {
+        MainScope.launch(handler) { executable.execute() }
+    }
 }
+
+internal actual inline fun OnEvent.Executor.isImmediate(): Boolean = this is OnEvent.Executor.Immediate
